[ADD] more test for ecs_vector

+++ include/utils/ecs_vector.h
@@ -40,32 +40,32 @@ ecs_vector3i_t *div_vector3i(ecs_vector3i_t *a, ecs_vector3i_t *b);
 ecs_vector3i_t *div_vector3i_value(ecs_vector3i_t *a, int value);
 
 typedef struct ecs_vector2f_s {
-    float x;
-    float y;
+    double x;
+    double y;
 } ecs_vector2f_t;
 
-void set_vector2f(ecs_vector2f_t *vector, float x, float y);
+void set_vector2f(ecs_vector2f_t *vector, double x, double y);
 ecs_vector2f_t *add_vector2f(ecs_vector2f_t *a, ecs_vector2f_t *b);
-ecs_vector2f_t *add_vector2f_value(ecs_vector2f_t *a, float value);
+ecs_vector2f_t *add_vector2f_value(ecs_vector2f_t *a, double value);
 ecs_vector2f_t *sub_vector2f(ecs_vector2f_t *a, ecs_vector2f_t *b);
-ecs_vector2f_t *sub_vector2f_value(ecs_vector2f_t *a, float value);
+ecs_vector2f_t *sub_vector2f_value(ecs_vector2f_t *a, double value);
 ecs_vector2f_t *mul_vector2f(ecs_vector2f_t *a, ecs_vector2f_t *b);
-ecs_vector2f_t *mul_vector2f_value(ecs_vector2f_t *a, float value);
+ecs_vector2f_t *mul_vector2f_value(ecs_vector2f_t *a, double value);
 ecs_vector2f_t *div_vector2f(ecs_vector2f_t *a, ecs_vector2f_t *b);
-ecs_vector2f_t *div_vector2f_value(ecs_vector2f_t *a, float value);
+ecs_vector2f_t *div_vector2f_value(ecs_vector2f_t *a, double value);
 
 typedef struct ecs_vector3f_s {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 } ecs_vector3f_t;
 
-void set_vector3f(ecs_vector3f_t *vector, float x, float y, float z);
+void set_vector3f(ecs_vector3f_t *vector, double x, double y, double z);
 ecs_vector3f_t *add_vector3f(ecs_vector3f_t *a, ecs_vector3f_t *b);
-ecs_vector3f_t *add_vector3f_value(ecs_vector3f_t *a, float value);
+ecs_vector3f_t *add_vector3f_value(ecs_vector3f_t *a, double value);
 ecs_vector3f_t *sub_vector3f(ecs_vector3f_t *a, ecs_vector3f_t *b);
-ecs_vector3f_t *sub_vector3f_value(ecs_vector3f_t *a, float value);
+ecs_vector3f_t *sub_vector3f_value(ecs_vector3f_t *a, double value);
 ecs_vector3f_t *mul_vector3f(ecs_vector3f_t *a, ecs_vector3f_t *b);
-ecs_vector3f_t *mul_vector3f_value(ecs_vector3f_t *a, float value);
+ecs_vector3f_t *mul_vector3f_value(ecs_vector3f_t *a, double value);
 ecs_vector3f_t *div_vector3f(ecs_vector3f_t *a, ecs_vector3f_t *b);
-ecs_vector3f_t *div_vector3f_value(ecs_vector3f_t *a, float value);
+ecs_vector3f_t *div_vector3f_value(ecs_vector3f_t *a, double value);
